move scheduled actions of game to rooms
